# GestionDeProductos/swaggerjson/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GestionDeProductos.Api",
    "version": "1.0"
  },
  "paths": {
    "/Deposito/GetDeposito": {
      "get": {
        "tags": [
          "Deposito"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Deposito"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deposito"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deposito"
                }
              }
            }
          }
        }
      }
    },
    "/Deposito/GetAllDepositos": {
      "get": {
        "tags": [
          "Deposito"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Deposito"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Deposito"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Deposito"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Deposito/CreateDeposito": {
      "post": {
        "tags": [
          "Deposito"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Deposito"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Deposito"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Deposito"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Deposito/UpdateDeposito": {
      "put": {
        "tags": [
          "Deposito"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Deposito"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Deposito"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Deposito"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Deposito/DeleteDeposito": {
      "delete": {
        "tags": [
          "Deposito"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Deposito/GetAllProductoDeposito": {
      "post": {
        "tags": [
          "Deposito"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductoDeposito"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductoDeposito"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductoDeposito"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Deposito/GetProductoDeposito": {
      "post": {
        "tags": [
          "Deposito"
        ],
        "parameters": [
          {
            "name": "idDeposito",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idProducto",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoDeposito"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoDeposito"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoDeposito"
                }
              }
            }
          }
        }
      }
    },
    "/Deposito/AgregarProducto": {
      "post": {
        "tags": [
          "Deposito"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoDeposito"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoDeposito"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoDeposito"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Deposito/TransferirProductoADeposito": {
      "post": {
        "tags": [
          "Deposito"
        ],
        "parameters": [
          {
            "name": "idDeposito",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "cantidad",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoDeposito"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoDeposito"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoDeposito"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Deposito/TransferirProductoATienda": {
      "post": {
        "tags": [
          "Deposito"
        ],
        "parameters": [
          {
            "name": "idTienda",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "cantidad",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoDeposito"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoDeposito"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoDeposito"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Operacion/GetOperacion": {
      "get": {
        "tags": [
          "Operacion"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Operacion"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Operacion"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Operacion"
                }
              }
            }
          }
        }
      }
    },
    "/Operacion/GetAllOperacions": {
      "get": {
        "tags": [
          "Operacion"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Operacion"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Operacion"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Operacion"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Operacion/CreateOperacion": {
      "post": {
        "tags": [
          "Operacion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Operacion"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Operacion"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Operacion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Operacion/UpdateOperacion": {
      "put": {
        "tags": [
          "Operacion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Operacion"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Operacion"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Operacion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Operacion/DeleteOperacion": {
      "delete": {
        "tags": [
          "Operacion"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Product/GetProduct": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Producto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Producto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Producto"
                }
              }
            }
          }
        }
      }
    },
    "/Product/GetAllProducts": {
      "get": {
        "tags": [
          "Product"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Producto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Producto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Producto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Product/CreateProduct": {
      "post": {
        "tags": [
          "Product"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Product/UpdateProduct": {
      "put": {
        "tags": [
          "Product"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Product/DeleteProduct": {
      "delete": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Tienda/GetTienda": {
      "get": {
        "tags": [
          "Tienda"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Tienda"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tienda"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tienda"
                }
              }
            }
          }
        }
      }
    },
    "/Tienda/GetAllTiendas": {
      "get": {
        "tags": [
          "Tienda"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tienda"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tienda"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tienda"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Tienda/CreateTienda": {
      "post": {
        "tags": [
          "Tienda"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tienda"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Tienda"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Tienda"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Tienda/UpdateTienda": {
      "put": {
        "tags": [
          "Tienda"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tienda"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Tienda"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Tienda"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Tienda/DeleteTienda": {
      "delete": {
        "tags": [
          "Tienda"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Deposito": {
        "type": "object",
        "properties": {
          "idDeposito": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "provincia": {
            "type": "string",
            "nullable": true
          },
          "localidad": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Operacion": {
        "type": "object",
        "properties": {
          "idOperacion": {
            "type": "integer",
            "format": "int32"
          },
          "origen": {
            "type": "integer",
            "format": "int32"
          },
          "destino": {
            "type": "integer",
            "format": "int32"
          },
          "esDeposito": {
            "type": "boolean"
          },
          "fecha": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "usuario": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Producto": {
        "type": "object",
        "properties": {
          "idProducto": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ProductoDeposito": {
        "type": "object",
        "properties": {
          "idDeposito": {
            "type": "integer",
            "format": "int32"
          },
          "idProducto": {
            "type": "integer",
            "format": "int32"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Tienda": {
        "type": "object",
        "properties": {
          "idtienda": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "provincia": {
            "type": "string",
            "nullable": true
          },
          "localidad": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}